+++ LUDAMAS/src/fr/irit/smac/planification/datadisplay/ui/DataAgentDisplay.java
@@ -1,4 +1,4 @@
-package fr.irit.smac.planification.datadisplayui;
+package fr.irit.smac.planification.datadisplay.ui;
 
 import javafx.geometry.Insets;
 import javafx.geometry.Orientation;
