+++ LUDAMAS/src/fr/irit/smac/planification/datadisplay/ui/OpenChartDisplayHandler.java
@@ -1,4 +1,4 @@
-package fr.irit.smac.planification.datadisplayui;
+package fr.irit.smac.planification.datadisplay.ui;
 
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
