+++ LUDAMAS/src/fr/irit/smac/planification/datadisplay/ui/OpenDataDisplayHandler.java
@@ -1,13 +1,13 @@
-package fr.irit.smac.planification.datadisplayui;
+package fr.irit.smac.planification.datadisplay.ui;
 
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 
-public class OpenDataMorphDisplayHandler implements EventHandler<ActionEvent> {
+public class OpenDataDisplayHandler implements EventHandler<ActionEvent> {
 
 	public void handle(ActionEvent actionEvent) {
 		DataAgentDisplay dataAgentDisplay = new DataAgentDisplay();
-		dataAgentDisplay.buildWindow(2);
+		dataAgentDisplay.buildWindow(1);
 	}
 
 }
