+++ LUDAMAS/src/fr/irit/smac/planification/datadisplay/ui/OpenDataMorphDisplayHandler.java
@@ -1,12 +1,13 @@
-package fr.irit.smac.planification.datadisplayui;
+package fr.irit.smac.planification.datadisplay.ui;
 
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 
-public class OpenEffectorDisplayHandler implements EventHandler<ActionEvent> {
+public class OpenDataMorphDisplayHandler implements EventHandler<ActionEvent> {
 
 	public void handle(ActionEvent actionEvent) {
 		DataAgentDisplay dataAgentDisplay = new DataAgentDisplay();
-		dataAgentDisplay.buildWindow(3);
+		dataAgentDisplay.buildWindow(2);
 	}
+
 }
