+++ LUDAMAS/src/fr/irit/smac/planification/datadisplay/ui/OpenEffectorDisplayHandler.java
@@ -1,13 +1,12 @@
-package fr.irit.smac.planification.datadisplayui;
+package fr.irit.smac.planification.datadisplay.ui;
 
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 
-public class OpenDataDisplayHandler implements EventHandler<ActionEvent> {
+public class OpenEffectorDisplayHandler implements EventHandler<ActionEvent> {
 
 	public void handle(ActionEvent actionEvent) {
 		DataAgentDisplay dataAgentDisplay = new DataAgentDisplay();
-		dataAgentDisplay.buildWindow(1);
+		dataAgentDisplay.buildWindow(3);
 	}
-
 }
